batchfix: get a *clean* path, not an absolute path
Fixes a failure when the current working directory isn't the same as the
batch base path. filepath.Abs is the *wrong* method for the intended
functionality here, because it uses the current dir if the path isn't
absolute, and we pass in an afero.Fs which is already pinned to the base
batch directory.